LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-lanes/track-sizing/auto-repeat/intrinsic-auto-repeat/column-auto-repeat-auto-021-expected.txt


FAIL Auto-fit repeat tracks with auto size and auto item placement assert_equals: expected "100px 100px 100px 100px 100px 100px 100px 100px 100px 100px" but got "1000px"
